Four classes? Two classes!
*I just went from having `abstract class EventType`, an `interface EventHandler<? extends EventType>`, an `abstract class EventData<? extends EventType>` and a `class Event<? extends EventType>`, to JUST an **empty**/marker `interface Event` and a functional interface `interface EventHandler<? extends Event>`.*

Why? Because we just decided that having a single EDT is not the way how everything should go!

`EventData` could simply be merged into `Event`!
...And `EventType` got eliminated because `instanceof` exists!

I plan to bring `EventType` back and make it a field in `Event` for its "single EDT for many events" flexibility - because an `interface EventType` could be implemented by an `enum` - which could declare many `EventType`s that could then use the same EDT - the one that's stored as a `static` field in the `enum` itself!